Create ADC Interface
