+++ basilika_pca10040.emProject
@@ -15,7 +15,7 @@
       arm_simulator_memory_simulation_parameter="RWX 00000000,00100000,FFFFFFFF;RWX 20000000,00010000,CDCDCDCD"
       arm_target_device_name="nRF52832_xxAA"
       arm_target_interface_type="SWD"
-      c_user_include_directories="config;nRF5_SDK_16.0.0_98a08e2/components;nRF5_SDK_16.0.0_98a08e2/components/boards;nRF5_SDK_16.0.0_98a08e2/components/drivers_nrf/nrf_soc_nosd;nRF5_SDK_16.0.0_98a08e2/components/libraries/atomic;nRF5_SDK_16.0.0_98a08e2/components/libraries/atomic_fifo;nRF5_SDK_16.0.0_98a08e2/components/libraries/balloc;nRF5_SDK_16.0.0_98a08e2/components/libraries/bsp;nRF5_SDK_16.0.0_98a08e2/components/libraries/button;nRF5_SDK_16.0.0_98a08e2/components/libraries/delay;nRF5_SDK_16.0.0_98a08e2/components/libraries/experimental_section_vars;nRF5_SDK_16.0.0_98a08e2/components/libraries/log;nRF5_SDK_16.0.0_98a08e2/components/libraries/log/src;nRF5_SDK_16.0.0_98a08e2/components/libraries/memobj;nRF5_SDK_16.0.0_98a08e2/components/libraries/ringbuf;nRF5_SDK_16.0.0_98a08e2/components/libraries/scheduler;nRF5_SDK_16.0.0_98a08e2/components/libraries/sortlist;nRF5_SDK_16.0.0_98a08e2/components/libraries/strerror;nRF5_SDK_16.0.0_98a08e2/components/libraries/timer;nRF5_SDK_16.0.0_98a08e2/components/libraries/util;nRF5_SDK_16.0.0_98a08e2/components/toolchain/cmsis/include;nRF5_SDK_16.0.0_98a08e2/external/fprintf;nRF5_SDK_16.0.0_98a08e2/external/segger_rtt;nRF5_SDK_16.0.0_98a08e2/integration/nrfx;nRF5_SDK_16.0.0_98a08e2/integration/nrfx/legacy;nRF5_SDK_16.0.0_98a08e2/modules/nrfx;nRF5_SDK_16.0.0_98a08e2/modules/nrfx/drivers/include;nRF5_SDK_16.0.0_98a08e2/modules/nrfx/hal;nRF5_SDK_16.0.0_98a08e2/modules/nrfx/mdk;nRF5_SDK_16.0.0_98a08e2/modules/nrfx;../config;"
+      c_user_include_directories="config;nRF5_SDK_16.0.0_98a08e2/components;nRF5_SDK_16.0.0_98a08e2/components/boards;nRF5_SDK_16.0.0_98a08e2/components/drivers_nrf/nrf_soc_nosd;nRF5_SDK_16.0.0_98a08e2/components/libraries/atomic;nRF5_SDK_16.0.0_98a08e2/components/libraries/atomic_fifo;nRF5_SDK_16.0.0_98a08e2/components/libraries/balloc;nRF5_SDK_16.0.0_98a08e2/components/libraries/bsp;nRF5_SDK_16.0.0_98a08e2/components/libraries/button;nRF5_SDK_16.0.0_98a08e2/components/libraries/delay;nRF5_SDK_16.0.0_98a08e2/components/libraries/experimental_section_vars;nRF5_SDK_16.0.0_98a08e2/components/libraries/log;nRF5_SDK_16.0.0_98a08e2/components/libraries/log/src;nRF5_SDK_16.0.0_98a08e2/components/libraries/memobj;nRF5_SDK_16.0.0_98a08e2/components/libraries/ringbuf;nRF5_SDK_16.0.0_98a08e2/components/libraries/scheduler;nRF5_SDK_16.0.0_98a08e2/components/libraries/sortlist;nRF5_SDK_16.0.0_98a08e2/components/libraries/strerror;nRF5_SDK_16.0.0_98a08e2/components/libraries/timer;nRF5_SDK_16.0.0_98a08e2/components/libraries/util;nRF5_SDK_16.0.0_98a08e2/components/toolchain/cmsis/include;nRF5_SDK_16.0.0_98a08e2/external/fprintf;nRF5_SDK_16.0.0_98a08e2/external/segger_rtt;nRF5_SDK_16.0.0_98a08e2/integration/nrfx;nRF5_SDK_16.0.0_98a08e2/integration/nrfx/legacy;nRF5_SDK_16.0.0_98a08e2/modules/nrfx;nRF5_SDK_16.0.0_98a08e2/modules/nrfx/drivers/include;nRF5_SDK_16.0.0_98a08e2/modules/nrfx/hal;nRF5_SDK_16.0.0_98a08e2/modules/nrfx/mdk;nRF5_SDK_16.0.0_98a08e2/modules/nrfx;../config;./interfaces;./interfaces/adc;"
       c_preprocessor_definitions="APP_TIMER_V2;APP_TIMER_V2_RTC1_ENABLED;BOARD_PCA10040;CONFIG_GPIO_AS_PINRESET;FLOAT_ABI_HARD;INITIALIZE_USER_SECTIONS;NO_VTOR_CONFIG;NRF52;NRF52832_XXAA;NRF52_PAN_74;"
       debug_target_connection="J-Link"
       gcc_entry_point="Reset_Handler"
@@ -97,6 +97,12 @@
       <file file_name="nRF5_SDK_16.0.0_98a08e2/modules/nrfx/mdk/ses_startup_nrf_common.s" />
       <file file_name="nRF5_SDK_16.0.0_98a08e2/modules/nrfx/mdk/system_nrf52.c" />
     </folder>
+    <folder 
+      Name="Interfaces"
+      exclude=""
+      filter="*.c;*.h"
+      path="interfaces"
+      recurse="Yes" />
   </project>
   <configuration Name="Release"
     c_preprocessor_definitions="NDEBUG"
