+++ interfaces/adc/saadc.h
@@ -1,0 +1,9 @@
+#ifndef SAADC_H
+#define SAADC_H
+
+void saadc_init(void);
+void saadc_uninit(void);
+void saadc_sample(void);
+void saadc_init_sample_uninit(void);
+
+#endif
